move EmailForm into a component, update readme.
